+++ src/main/java/http_server/ConfigurationRoutes.java
@@ -3,10 +3,10 @@ package http_server;
 import http_server.handler.HelloWorldGetHandler;
 import http_server.handler.directory.DirectoryGetHandler;
 import http_server.handler.directory.DirectoryHeadHandler;
-import http_server.handler.FormGetHandler;
-import http_server.handler.FormPostHandler;
-import http_server.handler.FormPutHandler;
-import http_server.handler.FormDeleteHandler;
+import http_server.handler.form.FormGetHandler;
+import http_server.handler.form.FormPostHandler;
+import http_server.handler.form.FormPutHandler;
+import http_server.handler.form.FormDeleteHandler;
 import http_server.handler.TeapotGetHandler;
 import http_server.handler.RedirectGetHandler;
 import http_server.handler.MethodsHandler;
